+++ src/Search/PostQueryBuilder.php
@@ -310,12 +310,31 @@ class PostQueryBuilder extends QueryBuilder {
     }
 
     /**
+     * WP_Query author parameter.
+     *
+     * @return string Redisearch query condition.
+     */
+    protected function author() : string {
+        if ( empty( $this->query->query_vars['author'] ) ) {
+            return false;
+        }
+
+        $author = $this->query->query_vars['author'];
+        $clause     = '';
+
+        if ( ! empty( $author ) && is_string( $author ) ) {
+            $clause = '@post_author_id:(' . $author . ')';
+        }
+
+        return $clause;
+    }
+
+    /**
      * WP_Query author__in parameter.
      *
      * @return string Redisearch query condition.
      */
     protected function author__in() : string {
-
         if ( empty( $this->query->query_vars['author__in'] ) ) {
             return false;
         }
@@ -324,7 +343,7 @@ class PostQueryBuilder extends QueryBuilder {
         $clause     = '';
 
         if ( ! empty( $author__in ) && is_array( $author__in ) ) {
-            $clause = '@post_author:(' . implode( '|', $author__in ) . ')';
+            $clause = '@post_author_id:(' . implode( '|', $author__in ) . ')';
         }
 
         return $clause;
@@ -336,7 +355,6 @@ class PostQueryBuilder extends QueryBuilder {
      * @return string Redisearch query condition.
      */
     protected function author__not_in() : string {
-
         if ( empty( $this->query->query_vars['author__not_in'] ) ) {
             return false;
         }
@@ -345,7 +363,7 @@ class PostQueryBuilder extends QueryBuilder {
         $clause         = '';
 
         if ( ! empty( $author__not_in ) && is_array( $author__not_in ) ) {
-            $clause = '-@post_author:(' . implode( '|', $author__not_in ) . ')';
+            $clause = '-@post_author_id:(' . implode( '|', $author__not_in ) . ')';
         }
 
         return $clause;
@@ -357,7 +375,6 @@ class PostQueryBuilder extends QueryBuilder {
      * @return ?string
      */
     protected function name() : ?string {
-
         if ( empty( $this->query->query_vars['name'] ) ) {
             return false;
         }